# services/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the account service for PayGidi.",
        "title": "PayGidi Account Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api.paygidi.site",
    "basePath": "/api/v1",
    "paths": {
        "/auth": {
            "post": {
                "description": "Initiate authentication (registration or login) via phone number. Sends an OTP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Initiate authentication",
                "parameters": [
                    {
                        "description": "Authentication data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.AuthDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "User not found (if accountType missing)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "User already exists (if accountType provided)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "429": {
                        "description": "Too many OTP requests",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/biometric": {
            "post": {
                "description": "Exchange a biometric ID for a new JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate via biometrics",
                "parameters": [
                    {
                        "description": "Biometric authentication data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.BiometricAuthDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/biometric/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Link a biometric ID to the authenticated user's account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Enable biometric authentication",
                "parameters": [
                    {
                        "description": "Biometric registration data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.RegisterBiometricDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Biometrics enabled successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/auth/verify": {
            "post": {
                "description": "Verify the OTP sent to the user's phone number and return JWT tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify authentication OTP",
                "parameters": [
                    {
                        "description": "OTP verification data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.VerifyAuthOtpDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid OTP or expired",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "OTP or User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "validators.AuthDto": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "accountType": {
                    "type": "string",
                    "enum": [
                        "individual",
                        "business"
                    ]
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                }
            }
        },
        "validators.BiometricAuthDto": {
            "type": "object",
            "required": [
                "biometricID",
                "phone"
            ],
            "properties": {
                "biometricID": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                }
            }
        },
        "validators.RegisterBiometricDto": {
            "type": "object",
            "required": [
                "biometricID"
            ],
            "properties": {
                "biometricID": {
                    "type": "string"
                }
            }
        },
...
        "validators.VerifyAuthOtpDto": {
            "type": "object",
            "required": [
                "otp",
                "phone"
            ],
            "properties": {
                "otp": {
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                }
            }
        },
...
    }
}